util: Change the type of stringref_t lengths to int.
And that's why everyone else uses `int` for these types, because it's
way more important to be able to easily subtract them from one another
and get negative values than it is to get 4 GiB rather than 2.

`ssize_t` would have been even nicer, but it's unsupported by Visual
Studio.

+++ thcrap/src/util.h
@@ -40,13 +40,13 @@ __inline char* strncpy_advance_dst(char *dst, const char *src, size_t len)
 // TODO: Rip out and change to std::string_view once C++17 is more widespread.
 typedef struct stringref_t {
 	const char *str;
-	size_t len;
+	int len;
 
 	// No default constructor = no potential uninitialized
 	// string pointer = good
 
 	stringref_t(const char *str) : str(str), len(strlen(str)) {}
-	stringref_t(const char *str, size_t len) : str(str), len(len) {}
+	stringref_t(const char *str, int len) : str(str), len(len) {}
 
 	stringref_t(const json_t *json)
 		: str(json_string_value(json)), len(json_string_length(json)) {
